+++ src/ast/number-exp.hxx
@@ -3,7 +3,7 @@
 #include "number-exp.hh"
 
 namespace ast {
-  inline int NumberExp::value_get() const {
+  inline float NumberExp::value_get() const {
     return value_;
   }
 } // namespace ast
