+++ src/parse/dreamberd-driver.hxx
@@ -3,7 +3,7 @@
 #include "dreamberd-driver.hh"
 
 namespace parse {
-  inline ast::NumberExp* DreamBerdDriver::make_NumberExp(const parse::location& location, int value) const {
+  inline ast::NumberExp* DreamBerdDriver::make_NumberExp(const parse::location& location, float value) const {
     return new ast::NumberExp(location, value);
   }
